# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DOG DATA API",
    "description": "Comprehensive on-chain analytics API for DOG (Bitcoin Rune 840000:3). Provides real-time holder data, transaction analytics, UTXO metrics, forensic behavioral analysis, airdrop tracking, market data from 20+ exchanges, and Bitcoin network stats. Designed for AI agents and programmatic access.",
    "version": "1.0.0",
    "contact": {
      "name": "DOG DATA",
      "url": "https://dogdata.xyz"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.dogdata.xyz/v1",
      "description": "Production API (v1 prefix)"
    },
    {
      "url": "https://www.dogdata.xyz/api",
      "description": "Production (legacy, direct routes)"
    }
  ],
  "security": [],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Optional bearer token for authenticated access and higher rate limits"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "status": { "type": "integer" },
              "retry_after": { "type": "integer", "description": "Seconds to wait before retrying (rate limit)" },
              "docs_url": { "type": "string", "format": "uri" }
            },
            "required": ["code", "message", "status"]
          }
        }
      },
...
    },
...
    "responses": {
...
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": { "code": "RATE_LIMITED", "message": "Too many requests", "status": 429, "retry_after": 60 }
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": { "code": "INTERNAL_ERROR", "message": "An unexpected error occurred", "status": 500 }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/multichain/holders": {
      "get": {
        "operationId": "getMultichainHolders",
        "tags": ["Multichain"],
        "summary": "Get DOG holders on Stacks and/or Solana",
        "description": "Returns top holders of bridged DOG tokens on Stacks (via Tenero) and Solana (via Helius). Includes balances, USD values, holder stats, and concentration percentages. Omit `chain` to get both chains.",
        "parameters": [
          {
            "name": "chain",
            "in": "query",
            "description": "Filter by blockchain: stacks or solana. Omit for both.",
            "required": false,
            "schema": { "type": "string", "enum": ["stacks", "solana"] }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of top holders to return (default 20)",
            "required": false,
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Multi-chain holder data with stats and concentration metrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stacks": {
                      "type": "object",
                      "description": "Stacks holders, stats, and concentration data"
                    },
                    "solana": {
                      "type": "object",
                      "description": "Solana holders with USD-enriched balances"
                    }
                  }
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/v1/multichain/transactions": {
      "get": {
        "operationId": "getMultichainTransactions",
        "tags": ["Multichain"],
        "summary": "Get DOG transactions on Stacks and/or Solana",
        "description": "Returns recent DOG token transactions (transfers and swaps) on Stacks (via Tenero) and Solana (via Helius Enhanced API). Omit `chain` to get both chains. Gracefully degrades if one chain is unavailable.",
        "parameters": [
          {
            "name": "chain",
            "in": "query",
            "description": "Filter by blockchain: stacks or solana. Omit for both.",
            "required": false,
            "schema": { "type": "string", "enum": ["stacks", "solana"] }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of transactions to return (default 30)",
            "required": false,
            "schema": { "type": "integer", "default": 30 }
          }
        ],
        "responses": {
          "200": {
            "description": "Multi-chain transaction feeds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stacks": {
                      "type": "object",
                      "description": "Stacks transactions (trades + transfers)"
                    },
                    "solana": {
                      "type": "object",
                      "description": "Solana DOG token transfers and swaps"
                    }
                  }
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  }
}